🐛id 타입 수정(Int -> Long)

+++ app/src/main/java/com/project/heyboardgame/adapter/SignUpRVAdapter.kt
@@ -24,8 +24,8 @@ class SignUpRVAdapter(private val genreList : MutableList<SignUpItem>)
     }
 
     // 선택된 장르들의 ID를 반환하는 함수
-    fun getSelectedItems(): List<Int> {
-        val selectedItems = mutableListOf<Int>()
+    fun getSelectedItems(): List<Long> {
+        val selectedItems = mutableListOf<Long>()
         for (item in genreList) {
             if (item.isClicked) {
                 selectedItems.add(item.getItemId())
