+++ app/src/main/java/com/project/heyboardgame/dataModel/SignUpData.kt
@@ -13,6 +13,6 @@ data class SignUpData(
     @SerializedName("nickname")
     val nickname : String,
     @SerializedName("favoriteGenreIds")
-    val id : List<Int>
+    val id : List<Long>
 ) : Parcelable
