+++ app/src/main/java/com/project/heyboardgame/dataModel/SignUpItem.kt
@@ -3,10 +3,10 @@ package com.project.heyboardgame.dataModel
 data class SignUpItem(
     val genreIcon: Int,
     val genreName: String,
-    val genreId: Int,
+    val genreId: Long,
     var isClicked: Boolean
 ) {
-    fun getItemId(): Int {
+    fun getItemId(): Long {
         return genreId
     }
 }
